requirements.txt
beautifulsoup4==4.8.1
lxml==4.9.1
PyYAML==6.0.2
requests==2.32.0
selenium==3.141.0
urllib3<2
